Update the configs to match what is called in the Provider
Provider.php calls for keys `site` and `key` from the config. The `key` id different from the `client_id` and it needs both. When you go https://stackapps.com/apps/oauth/register it issues an Id, Key, and Secret. 

The site needs to be one of their providers: `stackoverflow` works.